+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-08T13:40:53.886Z",
+    "2026-04-08T15:50:15.158Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {