+++ packages/orq-rc/RELEASES.md
@@ -8129,3 +8129,13 @@ Based on:
 - [typescript v4.7.0-rc.56] packages/orq-rc
 ### Releases
 - [NPM v4.7.0-rc.56] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.56 - packages/orq-rc
+
+## 2026-04-08 15:54:16
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.1 (2.879.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.7.0-rc.57] packages/orq-rc
+### Releases
+- [NPM v4.7.0-rc.57] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.57 - packages/orq-rc